Added basic AI link to homepage.
